ReleaseNeededPlugin introduced #84
