+++ src/main/resources/META-INF/gradle-plugins/org.mockito.mockito-release-tools.release-needed.properties
@@ -1,0 +1,1 @@
+implementation-class=org.mockito.release.internal.gradle.ReleaseNeededPlugin
