+++ src/test/groovy/org/mockito/release/internal/gradle/ReleaseNeededPluginTest.groovy
@@ -1,0 +1,14 @@
+package org.mockito.release.internal.gradle
+
+import org.gradle.testfixtures.ProjectBuilder
+import spock.lang.Specification
+
+class ReleaseNeededPluginTest extends Specification {
+
+    def project = new ProjectBuilder().build()
+
+    def "applies"() {
+        expect:
+        project.plugins.apply("org.mockito.mockito-release-tools.release-needed")
+    }
+}
